week_6/application/pyproject.toml
[tool.poetry]
name = "kafka_app"
version = "0.1.0"
description = ""
authors = ["vuvuzella <j.tabacii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "1.4.6"
confluent-kafka = "^2.0.2"
requests = "^2.28.2"
httpx = "^0.23.3"
avro = "^1.11.1"
faust = "^1.10.4"
fastavro = "^1.7.3"
pydantic = "^1.10.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
kafka_app = "kafka_app.main:main"
